refactor: move piano roll related hotkeys to usePianoRollKeyboardShortcut
