+++ app/src/hooks/usePianoRollKeyboardShortcut.tsx
@@ -1,5 +1,12 @@
 import { useMemo } from "react"
-import { useSelectAllNotes } from "../actions"
+import {
+  useNextTrack,
+  usePreviousTrack,
+  useSelectAllNotes,
+  useToggleGhost,
+  useToggleMute,
+  useToggleSolo,
+} from "../actions"
 import { useKeyboardShortcut } from "./useKeyboardShortcut"
 import { usePianoRoll } from "./usePianoRoll"
 
@@ -8,6 +15,11 @@ const SCROLL_DELTA = 24
 export const usePianoRollKeyboardShortcut = () => {
   const { setMouseMode, scrollBy } = usePianoRoll()
   const selectAllNotes = useSelectAllNotes()
+  const nextTrack = useNextTrack()
+  const previousTrack = usePreviousTrack()
+  const toggleSolo = useToggleSolo()
+  const toggleMute = useToggleMute()
+  const toggleGhost = useToggleGhost()
 
   const actions = useMemo(
     () => [
@@ -44,8 +56,27 @@ export const usePianoRollKeyboardShortcut = () => {
         metaKey: true,
         run: selectAllNotes,
       },
+      // Next track (S)
+      { code: "KeyS", run: nextTrack },
+      // Previous track (W)
+      { code: "KeyW", run: previousTrack },
+      // Toggle solo (N)
+      { code: "KeyN", run: toggleSolo },
+      // Toggle mute (M)
+      { code: "KeyM", run: toggleMute },
+      // Toggle ghost (,)
+      { code: "Comma", run: toggleGhost },
     ],
-    [scrollBy, setMouseMode, selectAllNotes],
+    [
+      scrollBy,
+      setMouseMode,
+      selectAllNotes,
+      nextTrack,
+      previousTrack,
+      toggleSolo,
+      toggleMute,
+      toggleGhost,
+    ],
   )
 
   return useKeyboardShortcut({
